✂ Split texture data dependency on System.Drawing.Bitmap into separate project
Font loading still depends on System.Drawing to render the texture atlas internally, but we'll handle that separately another time
